Added some more tests. Tentatively added jcip annotations.
git-svn-id: https://sulky.svn.sourceforge.net/svnroot/sulky/trunk@78 33f67c6c-4047-0410-b7fd-cc5e57a77e41

+++ sulky-tasks/src/main/java/de/huxhorn/sulky/tasks/AbstractProgressingCallable.java
@@ -60,8 +60,8 @@ public abstract class AbstractProgressingCallable<T>
 	private int progress;
 	private long numberOfSteps;
 	private long currentStep;
-	private int initialSleepSteps;
-	private int laterSleepSteps;
+	private long initialSleepSteps;
+	private long laterSleepSteps;
 	private long lastSleepStep;
 
 	public AbstractProgressingCallable()
@@ -69,12 +69,12 @@ public abstract class AbstractProgressingCallable<T>
 		this(0, 0);
 	}
 
-	public AbstractProgressingCallable(int sleepSteps)
+	public AbstractProgressingCallable(long sleepSteps)
 	{
 		this(sleepSteps, 0);
 	}
 
-	public AbstractProgressingCallable(int initialSleepSteps, int laterSleepSteps)
+	public AbstractProgressingCallable(long initialSleepSteps, long laterSleepSteps)
 	{
 		this.changeSupport = new PropertyChangeSupport(this);
 		this.progress = -1;
